Updated index and style
Index.html: I added the links for the google fonts.

Style.css: I added the fonts to the ::root to be applied to everything. I also changed the height of the main element to fit-content due to the main container not adjusting the way I wanted it to. I also adjusted some settings for the fonts to hopefully be closer to the challenge.